Move tests in test_enkf_library
