+++ tests/libres_tests/res/enkf/data/test_enkf_config_node.py
@@ -3,7 +3,13 @@ import pytest
 from ert._c_wrappers.enkf.config import EnkfConfigNode
 
 
+def test_failed_enkf_config_node_creation():
+    with pytest.raises(NotImplementedError):
+        EnkfConfigNode()
+
+
 def test_gen_data():
+
     # Must have %d in filename argument
     with pytest.raises(ValueError):
         config_node = EnkfConfigNode.create_gen_data("KEY", "FILE")
